refactor: move files for better drectory structure
