+++ src/NvimClient.APIGenerator/Properties/Models/CSFunction.cs
@@ -1,5 +1,5 @@
-using NvimClient.NvimMsgpack;
-using NvimClient.NvimMsgpack.Models;
+using NvimClient.Models.MsgPack;
+using NvimClient.Models.Nvim;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
